added confirmation modal to package uninstall

+++ frontend/src/routes/packages/[id]/(components)/action/action-remove.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { getToastStore, type ToastSettings } from '@skeletonlabs/skeleton';
+    import { getToastStore, getModalStore, type ToastSettings, type ModalSettings } from '@skeletonlabs/skeleton';
 
     import { UninstallPackage } from '$lib/wailsjs/go/application/Driver';
     import { application } from '$lib/wailsjs/go/models';
@@ -7,6 +7,8 @@
     import { AlertAction } from '$lib/components/ui/alert';
 
     const toastStore = getToastStore();
+    const modalStore = getModalStore();
+
 
     export let packageId: string;
     export let cancel: boolean = false;
@@ -22,18 +24,36 @@
             return;
         }
 
-        disabled = true;
-        const opts = application.UninstallPackageOpts.createFrom({ id: packageId });
-
-        UninstallPackage(opts).then(() => {
-            return;
-        }).catch(error => {
-            const downloadPackageToast: ToastSettings = {
-                message: `Error removing installation: ${error}`,
-                background: "variant-filled-error"
+        new Promise<boolean>((resolve) => {
+            const modal: ModalSettings = {
+                type: "confirm",
+                title: "Please Confirm",
+                body: "Are you sure you wish to delete the downloaded content for this package?",
+                response: (r: boolean) => {
+                    resolve(r);
+                }
             };
+            modalStore.trigger(modal);
+        }).then(async (remove: boolean) => {
+            if (!remove) {
+                return;
+            }
 
-            toastStore.trigger(downloadPackageToast);
+            disabled = true;
+            const opts = application.UninstallPackageOpts.createFrom({ id: packageId });
+
+            UninstallPackage(opts).then(() => {
+                return;
+            }).catch(error => {
+                const downloadPackageToast: ToastSettings = {
+                    message: `Error removing installation: ${error}`,
+                    background: "variant-filled-error"
+                };
+
+                toastStore.trigger(downloadPackageToast);
+            });
+
+            disabled = false;
         });
     }
